Delete CFG_DONE signal from FPGA
We've named one of the pins connected to the FPGA as CFG_DONE. The
signal has always been high on the FPGA side. From the very begging,
we have an agreement internally to use the WDT signal to know whether
the FPGA is configured and kicking or not. This signal has been
reserved but was left connected until now.

We now found a better use of the pin and we'd like to remove the use
of CFG_DONE all together.

This commit also removes the transition from FPGA_STATE_ACTIVE to
FPGA_STATE_CONFIG as it is unnecessary.