fpga: Add ERROR state
Add a new ERROR state.  Currently we only have one error condition,
the watchdog error.  Move to ERROR state when watchdog is failed.  Let
User decide what to do.

For this commit, we keep the current behavior:

 - Power up the FPGA after PoR.
 - Power cycle the FPGA when an error happened.

Initialization of activate_fpga is moved from the beginning of every
iteration to FPGA_STATE_POWER_OFF.  That's what it intended to do.

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>