fpga: Move mem_select to main
Move the selection of FPGA configuration memory to User.  The state
machine does what User tells it to do.  The variable is renamed to
config_memory.

The behaviour is slightly changed.  With the old behaivour, it
switches to the other configuration memory when the state machine find
a WDT error or User asks to shutdown the FPGA by assigning 0 to the
activate_fpga variable.

The new behaviour selects the other configuration memory when and only
when an error is detected.

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>