[DNM] timer: Add wait_for
This version of wait_for() has a flaw. Do NOT use it. Having static
variable does not allow this function to be used at different location
at the same time.  This should be fixed.

But, does it worth having a function like this?

    tmp = false;
    if (wait_for(&tmp, &from, wait)) {
        // time has passed.  Do some work.
    }

tmp and from should be initialized once. Thus, a calling site should
be like:

    if (!initialized) {
        initialized = true
        from = timer_get_ticks();
    }
    if (wait_for(from)) {
        // time has passed.  Do some work.
    }

The current version mistakenly captured these two variables.

If we were using signed int for the start tick variable, we can encode
"uninitialized" state using a negative value.  This reduces required
variables down to two from three.

    from = -1;
    if (wait_for(&from, wait)) {
        // time has passed.  Do some work.
    }

This is much better than the three variables version.

If we are allowed to use a "static" variable in the middle of the main
while loop, we can do

    static from = timer_get_ticks();
    if (wait_for(from, wait)) {
        // time has passed.  Do some work.
    }

This could be a good candidate.

wait_for() will wait for given ticks and return true.  It returns
false otherwise.  The current version only handle 16 bits ticks or
65535 ticks, which is, with 10 HZ, about 1.8 hours.

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>