trch.h: Set SPICAN_CS_B HIGH at initial
SPICAN_CS_B is active low signal to select the on-board SPI CAN chip.
Even the documentation says it is initialized to HIGH, as you can see
from d49efb9e5516bf165.  But it was initialized to LOW.

The signal has a pull-up resistor with VDD_3V3_SYS and spi_init(),
which is called from main() at the very beginning, sets it HIGH
anyway.  So even though we initialize it to LOW, there was only a bit
of time the signal is actually LOW.

Fix it by setting it to HIGH from the beginning.

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>