trch.h: Fix bugs related to #define directive
This commit fixes some defines used by the `LOW()` macro in
trch.h. The fact that `LOW()` takes an argument is only for visual
cues to show our intention to users. It unconditionally uses `0`
regardless of the given argument.

Thus, this change does not affect generating the final binary at all.

Signed-off-by: Daisuke Sasaki <daisuke@spacecubics.com>