Makefile: Disable Watch Dog Reset by default
If the watchdog reset by TRCH is not disabled during FPGA or software
update by `west flash`, a reset may occur in the middle of the update
process. The watchdog kick stops during `west flash` because the CPU
is halted as part of the flashing process.

Therefore, users are currently required to manually disable the
watchdog reset when updates, which makes the procedure complex.

In the future, the watchdog reset should be controllable via some
external trigger. However, for now, the watchdog reset will be
disabled by default.

Signed-off-by: Takuya Sasaki <takuya.sasaki@spacecubics.com>